requirements.txt
pdfplumber==0.7.6
PyMuPDF==1.22.5
streamlit==1.26.0
pandas==2.2.2
python-dateutil==2.8.2
reportlab==4.0.7